refactor: change pages data folder and create mock to be used on tests
